CARLANE Dataset Tools/CARLANE Labeler/requirements.txt
opencv-python
numpy
pillow
pygame